Adding in calculations for crop quality into Calculator.java

+++ src/test/java/PlayerTest.java
@@ -8,7 +8,7 @@ public class PlayerTest {
     @Test
     public void shouldNotAssignProfessionsBasedOnFarmingLevelInput() {
 //        Given
-        int farmingLevel = 0;
+        double farmingLevel = 0;
         int balance = 100;
 //        When
         Player player = new Player(farmingLevel, balance);
